requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-levenshtein>=0.25
